+++ src/components/LinkItems/Item.vue
@@ -9,7 +9,7 @@
       v-tooltip="getTooltipOptions()"
       rel="noopener noreferrer" tabindex="0"
       :id="`link-${id}`"
-      :style="`--open-icon: ${getUnicodeOpeningIcon()}; ${customStyles}`"
+      :style="`--open-icon: ${getUnicodeOpeningIcon()}; color: ${color};  ${customStyles}`"
     >
       <!-- Item Text -->
       <div :class="`tile-title  ${!icon? 'bounce no-icon': ''}`" :id="`tile-${id}`" >
@@ -452,7 +452,7 @@ export default {
     height: 2rem;
     padding-top: 4px;
     max-width: 14rem;
-    div img, div svg.missing-image {
+    div img {
       width: 2rem;
     }
     .tile-title {
@@ -473,7 +473,7 @@ export default {
     flex-direction: column;
     align-items: center;
     height: auto;
-    div img, div svg.missing-image {
+    div img {
       width: 2.5rem;
       margin-bottom: 0.25rem;
     }